Dhananjay Singh 25_04_2025 12_19_PM
25_04_2025 12_19_PM

+++ app/src/main/java/com/mykaimeal/planner/fragment/mainfragment/commonscreen/statistics/model/LinkGenerateModel.kt
@@ -17,8 +17,8 @@ data class StatisticsGraphModel(
 data class StatisticsGraphModelData(
     val graph_data: GraphData,
     val month: String,
-    val saving: Int,
-    val total_spent: Int
+    val saving: Double,
+    val total_spent: Double
 )
 
 data class GraphData(
